#9 Test what is a realistic rate of small FIX messages over longer period of time.
FixTransportAppTest was used.

The test was run for ~8 minutes, on localhost to detect code bottlenecks.
There were 100 000 000 messages sent an received
on a single thread, with an average rate of
~ 200 000 msg/s (200 000 sent + 200 000 received per second)

Scalability wise,when run with batches of 20 000 000 messages on 4 CPUs,
it achieved a rate of  ~700 000 msg/s

Most on-cpu time was spent on sun/nio/ch/FileDispatcherImpl.write0
and sun/nio/ch/FileDispatcherImpl.read0, and some on-cpu time was
spent on sun/nio/ch/EPoll.wait

All detected allocations were caused by sun/nio/ch/SelectorImpl.selectNow
(not optimized by the zero allocation replacement yet)

The were 500-2000 microsecond pauses observed every ~20 seconds on G1 GC,
as logged by -XX:+PrintGCDetails

+++ asynctcpacceptance/src/test/java/dev/squaremile/asynctcpacceptance/sampleapps/FixTransportAppTest.java
@@ -1,6 +1,6 @@
 package dev.squaremile.asynctcpacceptance.sampleapps;
 
-import org.agrona.collections.MutableInteger;
+import org.agrona.collections.MutableLong;
 import org.junit.jupiter.api.Test;
 
 import static org.assertj.core.api.Assertions.assertThat;
@@ -17,7 +17,7 @@ import static java.util.concurrent.locks.LockSupport.parkNanos;
 public class FixTransportAppTest
 {
     private static final int NUMBER_OF_CYCLES = 100_000;
-    private final MutableInteger messageCount = new MutableInteger();
+    private final MutableLong messageCount = new MutableLong();
 
     @Test
     void shouldExchangeMessages()
